refactoring vectors
